requirements.txt
antiorm==1.2.1
asgiref==3.8.1
db==0.1.1
Django==5.1.6
djangorestframework==3.15.2
pillow==11.1.0
sqlparse==0.5.3
tzdata==2025.1
